Some width reformatting, some refactoring
I changed how the width was setup so that the activity sections and the bulk of the activity pages can each stretch and shrink their content with the changing width of a browser. I also noticed that the exact same code was used in two hover psuedo classes and opted to convert them into a single utility class to improve adherence to DRY principles.